+++ client/components/Application/Sidebar.tsx
@@ -1,44 +1,135 @@
-import Box from '@mui/material/Box'
-import Browser from './Browser'
-import LowerMenu from './LowerMenu'
-import sidebarLogo from '../../assets/ODE_sidebar_logo.svg'
-import Button from '@mui/material/Button'
 import * as store from '@client/store'
+import TranslateIcon from '@mui/icons-material/Translate'
+import Box from '@mui/material/Box'
+import Button from '@mui/material/Button'
+import FormControl from '@mui/material/FormControl'
+import InputLabel from '@mui/material/InputLabel'
+import MenuItem from '@mui/material/MenuItem'
+import Select from '@mui/material/Select'
+import Stack from '@mui/material/Stack'
 import { useTranslation } from 'react-i18next'
+import sidebarLogo from '../../assets/ODE_sidebar_logo.svg'
+import reportIssueIcon from '../../assets/report_issue_icon.svg'
+import userGuideIcon from '../../assets/user_guide_icon.svg'
+import Browser from './Browser'
 
 export default function Sidebar() {
-  const { t } = useTranslation()
   return (
-    <Box
-      className="sidebar"
+    <Stack
       sx={{
-        display: 'flex',
-        flexDirection: 'column',
+        height: '100vh',
         borderRight: 'solid 1px #ddd',
       }}
     >
-      <Box sx={{ padding: '24px' }}>
-        {' '}
-        <img src={sidebarLogo} alt="" />
-      </Box>
-      <Button
-        variant="outlined"
-        sx={{
-          textTransform: 'none',
-          marginBottom: '22px',
-          mx: '24px',
-          '&:hover': {
-            color: 'white',
-            borderColor: (theme) => theme.palette.OKFNBlue.main,
-            backgroundColor: (theme) => theme.palette.OKFNBlue.main,
-          },
-        }}
-        onClick={() => store.openDialog('fileUpload')}
-      >
-        {t('upload-your-data')}
-      </Button>
+      <SidebarLogo />
+      <SidebarControls />
       <Browser />
-      <LowerMenu />
+      <SidebarMenu />
+    </Stack>
+  )
+}
+
+export function SidebarLogo() {
+  return (
+    <Box sx={{ padding: '24px' }}>
+      <img src={sidebarLogo} alt="" />
     </Box>
+  )
+}
+
+export function SidebarControls() {
+  const { t } = useTranslation()
+
+  return (
+    <Button
+      variant="outlined"
+      sx={{
+        textTransform: 'none',
+        mb: '22px',
+        mx: '24px',
+        '&:hover': {
+          color: 'white',
+          borderColor: (theme) => theme.palette.OKFNBlue.main,
+          backgroundColor: (theme) => theme.palette.OKFNBlue.main,
+        },
+      }}
+      onClick={() => store.openDialog('fileUpload')}
+    >
+      {t('upload-your-data')}
+    </Button>
+  )
+}
+
+function SidebarMenu() {
+  return (
+    <Stack
+      spacing={0}
+      useFlexGap
+      sx={{
+        alignItems: 'flex-start',
+        borderTop: '1px solid #E6E7EB',
+        padding: '20px',
+      }}
+    >
+      <UserGuideLink />
+      <ReportIssueLink />
+      <LanguageSelect />
+    </Stack>
+  )
+}
+
+function UserGuideLink() {
+  return (
+    <Button
+      href="https://opendataeditor.okfn.org/documentation/getting-started/"
+      target="_blank"
+      color="OKFNCoolGray"
+      sx={{ textTransform: 'none' }}
+      startIcon={<img src={userGuideIcon} alt="" />}
+    >
+      User guide
+    </Button>
+  )
+}
+
+function ReportIssueLink() {
+  return (
+    <Button
+      href="https://github.com/okfn/opendataeditor"
+      target="_blank"
+      color="OKFNCoolGray"
+      sx={{ textTransform: 'none' }}
+      startIcon={<img src={reportIssueIcon} alt="" />}
+    >
+      Report an issue
+    </Button>
+  )
+}
+
+function LanguageSelect() {
+  return (
+    <Stack
+      direction="row"
+      spacing={1}
+      sx={{
+        marginTop: 1,
+        color: 'OKFNCoolGray.main',
+        alignItems: 'center',
+      }}
+    >
+      <TranslateIcon sx={{ fontSize: '20px', marginLeft: '4px' }} />
+      <FormControl size="small">
+        <InputLabel id="language-select-label">Language</InputLabel>
+        <Select
+          label="Language"
+          labelId="language-select-label"
+          value={'en'}
+          onChange={console.log}
+        >
+          <MenuItem value="en">English</MenuItem>
+          <MenuItem value="spanish">Spanish</MenuItem>
+        </Select>
+      </FormControl>
+    </Stack>
   )
 }
